Разбил на модули code_processor
