+++ code_processor/domain/result.go
@@ -1,0 +1,7 @@
+package domain
+
+type Result struct {
+	TaskId string `json:"task_id"`
+	Stdout string `json:"stdout"`
+	Stderr string `json:"stderr"`
+}
