+++ code_processor/domain/task.go
@@ -1,0 +1,7 @@
+package domain
+
+type Task struct {
+	TaskId     string `json:"task_id"`
+	Translator string `json:"translator"`
+	Code       string `json:"code"`
+}
